+++ src/lib.rs
@@ -368,6 +368,7 @@ pub mod pmv;
 pub mod quant_matrix_extension;
 pub mod quantizer_scale;
 pub mod sequence_display_extension;
+pub mod sequence_display_order;
 pub mod sequence_extension;
 pub mod sequence_header;
 pub mod skipped_macroblock;
@@ -482,6 +483,7 @@ pub use quantizer_scale::{QuantizerScale, QUANTIZER_SCALE_MAX, QUANTIZER_SCALE_M
 pub use sequence_display_extension::{
     ColourDescription, SequenceDisplayExtension, VideoFormat, SEQUENCE_DISPLAY_EXTENSION_ID,
 };
+pub use sequence_display_order::{Requirement, SequenceDisplayOrderDriver};
 pub use sequence_extension::{
     ChromaFormat, Mpeg2Sequence, Mpeg2SequenceExtension, EXTENSION_START_CODE,
     SEQUENCE_EXTENSION_ID,